feat: menu card details refactored
