+++ order-management-be/src/api/controllers/order.controller.js
@@ -36,7 +36,23 @@ const getTableDetails = async (req, res) => {
     }
 };
 
+const getMenuDetails = async (req, res) => {
+    try {
+        const { hotelId } = req.params;
+        logger('debug', `Fetching hotel details for cutomer ${hotelId}`);
+
+        const result = await orderService.getMenuDetails(hotelId);
+        logger('info', 'Hotel details fetched successfully', { result });
+
+        return res.status(STATUS_CODE.OK).send(result);
+    } catch (error) {
+        logger('error', `Error occurred during fetching hotel details ${JSON.stringify(error)}`);
+        return res.status(error.code).send({ message: error.message });
+    }
+};
+
 export default {
     register,
-    getTableDetails
+    getTableDetails,
+    getMenuDetails
 };
