+++ order-management-be/src/api/models/order.model.js
@@ -1,0 +1,53 @@
+import { DataTypes } from 'sequelize';
+import { TABLES } from '../utils/common.js';
+
+const ORDER_STATUS = ['PENDING', 'SERVED', 'COMPLETED'];
+const orderModel = (sequelize) =>
+    sequelize.define(
+        TABLES.ORDER,
+        {
+            id: {
+                type: DataTypes.STRING,
+                primaryKey: true
+            },
+            menuId: {
+                type: DataTypes.STRING,
+                allowNull: false,
+                references: {
+                    model: TABLES.MENU,
+                    key: 'id'
+                }
+            },
+            customerId: {
+                type: DataTypes.STRING,
+                allowNull: false,
+                references: {
+                    model: TABLES.CUSTOMER,
+                    key: 'id'
+                }
+            },
+            price: {
+                type: DataTypes.INTEGER,
+                allowNull: false
+            },
+            quantity: {
+                type: DataTypes.INTEGER,
+                allowNull: false
+            },
+            status: {
+                type: DataTypes.ENUM,
+                allowNull: false,
+                values: ORDER_STATUS,
+                defaultValue: ORDER_STATUS[0]
+            },
+            deletedAt: {
+                type: DataTypes.DATE,
+                allowNull: true
+            }
+        },
+        {
+            paranoid: true
+        }
+    );
+
+export default orderModel;
