+++ order-management-be/src/api/routes/hotel.routes.js
@@ -8,6 +8,5 @@ const router = Router();
 router.put('/:id', authenticate, hotelController.update);
 router.route('/').all(authenticate, ownerAuthentication).post(hotelController.register).get(hotelController.list);
 router.delete('/:id', authenticate, ownerAuthentication, hotelController.remove);
-router.get('/:hotelId', hotelController.getDetails);
 
 export default router;
