+++ order-management-be/src/api/routes/order.routes.js
@@ -5,5 +5,6 @@ const router = Router();
 
 router.post('/customer', orderController.register);
 router.get('/table/:id', orderController.getTableDetails);
+router.get('/menu/:hotelId', orderController.getMenuDetails);
 
 export default router;
