+++ order-management-be/src/api/services/order.service.js
@@ -1,8 +1,10 @@
 import { v4 as uuidv4 } from 'uuid';
 import { db } from '../../config/database.js';
 import logger from '../../config/logger.js';
+import { MENU_STATUS } from '../models/menu.model.js';
 import { TABLE_STATUS } from '../models/table.model.js';
 import customerRepo from '../repositories/customer.repository.js';
+import hotelRepo from '../repositories/hotel.repository.js';
 import tableRepo from '../repositories/table.repository.js';
 import { CustomError, STATUS_CODE } from '../utils/common.js';
 
@@ -57,7 +59,116 @@ const getTableDetails = async (id) => {
     }
 };
 
+const getMenuCardFormatData = ({ id, name, categories }) => {
+    const types = {
+        cover: 'COVER',
+        category: 'CATEGORY',
+        title: 'TITLE',
+        item: 'MENU_ITEM'
+    };
+
+    const typeData = {
+        cover: [{ name, id }]
+    };
+
+    categories.forEach(({ id: categoryId, name, menus }) => {
+        if (!typeData.category) {
+            typeData.category = [];
+            typeData.menuData = {};
+        }
+
+        typeData.category.push({
+            name,
+            id: categoryId
+        });
+
+        const menuItemData = [];
+        menus.forEach((item) => {
+            menuItemData.push({
+                name: item.name,
+                id: item.id,
+                price: item.price
+            });
+        });
+        typeData.menuData = {
+            ...typeData.menuData,
+            [`${categoryId}_${name}`]: menuItemData
+        };
+    });
+
+    const data = {};
+    let page = 0;
+    data[page] = { type: types.cover, data: typeData.cover[0] };
+    page++;
+
+    const categoriesPerPage = 12;
+    const categoriesCount = Math.ceil(typeData.category.length / categoriesPerPage);
+    for (let index = page; index < page + categoriesCount; index++) {
+        data[index] = {
+            title: 'Categories',
+            type: types.category,
+            data: typeData.category.splice(0, categoriesPerPage)
+        };
+    }
+    page += categoriesCount;
+
+    const mapping = {};
+    const menusPerPage = 10;
+    Object.keys(typeData.menuData).forEach((key) => {
+        const id = key.split('_')[0];
+        const name = key.split('_')[1];
+        mapping[id] = page;
+        const menuCount = Math.ceil(typeData.menuData[key].length / menusPerPage);
+        const menus = typeData.menuData[key];
+        for (let index = page; index < page + menuCount; index++) {
+            data[index] = { title: name, type: types.item, data: menus.splice(0, menusPerPage) };
+        }
+        page += menuCount;
+    });
+
+    return { data, mapping };
+};
+
+const getMenuDetails = async (hotelId) => {
+    try {
+        const options = {
+            where: { id: hotelId },
+            attributes: ['id', 'name'],
+            include: [
+                {
+                    model: db.categories,
+                    attributes: ['id', 'name'],
+                    include: [
+                        {
+                            model: db.menu,
+                            where: { status: MENU_STATUS[0] },
+                            attributes: ['id', 'name', 'price']
+                        }
+                    ]
+                }
+            ],
+            order: [[db.categories, 'order', 'ASC']]
+        };
+        logger('debug', 'Fetching hotels details');
+
+        const res = await hotelRepo.find(options);
+        const { data: formatedData, mapping } = getMenuCardFormatData(res);
+
+        return {
+            id: res.id,
+            name: res.name,
+            count: Object.keys(formatedData).length,
+            data: formatedData,
+            mapping
+        };
+    } catch (error) {
+        logger('error', 'Error while detching hotel details', { error });
+        throw CustomError(error.code, error.message);
+    }
+};
+
 export default {
     register,
-    getTableDetails
+    getTableDetails,
+    getMenuDetails
 };
